requirements.txt
pandas
google-cloud-language
google-cloud-storage
fsspec
gcsfs